Lock fps to reduce cpu usage.

+++ Source/ShaderHelper/App/ShaderHelperApp.cpp
@@ -40,7 +40,7 @@ namespace SH {
 		}
 	}
 
-	void ShaderHelperApp::Update(double DeltaTime)
+	void ShaderHelperApp::Update(float DeltaTime)
 	{
 		App::Update(DeltaTime);
 	}
