+++ Source/Tests/UnitTestUtil/App/UnitTestApp.cpp
@@ -18,7 +18,7 @@ namespace UNITTEST_UTIL
 		AppEditor = MakeUnique<UnitTestEditor>(AppClientSize);
 	}
 
-	void UnitTestApp::Update(double DeltaTime)
+	void UnitTestApp::Update(float DeltaTime)
 	{
 		App::Update(DeltaTime);
 	}
